refactor(AttendanceRepository): 제적 위험자 조회 메서드 이동
기존 AttendanceBook에서 갖고있던 제적 위험자 조회 메서드를 순환참조 방지 및 책임 이전을 위해 AttendanceRepository로 이동합니다.
