+++ src/test/java/attendance/repository/AttendanceRepositoryTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import attendance.domain.Attendance;
+import attendance.domain.AttendanceBook;
 import attendance.domain.AttendanceTime;
 import attendance.domain.CrewAttendanceInformation;
 import java.time.DayOfWeek;
@@ -13,6 +14,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
@@ -20,6 +22,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class AttendanceRepositoryTest {
@@ -188,5 +191,49 @@ class AttendanceRepositoryTest {
 
         DayOfWeek dayOfWeek = date.getDayOfWeek();
         return dayOfWeek == DayOfWeek.SATURDAY || dayOfWeek == DayOfWeek.SUNDAY;
+    }
+
+    @DisplayName("제적 위험자 리스트를 가져온다.")
+    @ParameterizedTest
+    @CsvSource(value = {
+            "체체,제적", "체글렛,면담", "피글렛,경고"
+    })
+    void 제적_위험자_리스트를_가져온다(String name, String academicStatus) {
+
+        // given
+        AttendanceBook attendanceBook = new AttendanceBook(Set.of("체체", "체글렛", "피글렛", "피글체"));
+        List<Attendance> attendances = makeAttendance();
+        AttendanceRepository attendanceRepository = new AttendanceRepository(attendances);
+
+        // when
+        List<CrewAttendanceInformation> crewAttendanceInformations = attendanceRepository.getCrewAtRiskOfExpulsion(
+                attendanceBook.getNames(), academicStatus);
+
+        // then
+        assertAll(() -> {
+            assertEquals(crewAttendanceInformations.size(), 1);
+            assertEquals(crewAttendanceInformations.getFirst().crewName(), name);
+        });
+    }
+
+    private static List<Attendance> makeAttendance() {
+
+        List<Attendance> attendances = new ArrayList<>(List.of(makeAbsentAttendance("체체", 2025, 2, 10),
+                makeAbsentAttendance("체체", 2025, 2, 11),
+                makeAbsentAttendance("체체", 2025, 2, 12),
+                makeAbsentAttendance("체체", 2025, 2, 13),
+                makeAbsentAttendance("체체", 2025, 2, 14),
+                makeAbsentAttendance("체체", 2025, 2, 17)));
+
+        attendances.addAll(List.of(makeAbsentAttendance("피글렛", 2025, 2, 10),
+                makeAbsentAttendance("피글렛", 2025, 2, 11)));
+
+        attendances.addAll(List.of(makeAbsentAttendance("체글렛", 2025, 2, 10),
+                makeAbsentAttendance("체글렛", 2025, 2, 11),
+                makeAbsentAttendance("체글렛", 2025, 2, 12)));
+
+        attendances.add(makeAbsentAttendance("피글체", 2025, 2, 10));
+
+        return attendances;
     }
 }
